+++ chia/_tests/wallet/clawback/test_clawback_lifecycle.py
@@ -12,12 +12,12 @@ from chia._tests.clvm.test_puzzles import public_key_for_index, secret_exponent_
 from chia._tests.util.key_tool import KeyTool
 from chia._tests.util.spend_sim import CostLogger, SimClient, SpendSim, sim_and_client
 from chia._tests.util.time_out_assert import time_out_assert
+from chia.consensus.condition_tools import conditions_dict_for_solution, pkm_pairs_for_conditions_dict
 from chia.consensus.default_constants import DEFAULT_CONSTANTS
 from chia.types.blockchain_format.program import INFINITE_COST, Program
 from chia.types.coin_spend import make_spend
 from chia.types.condition_opcodes import ConditionOpcode
 from chia.types.mempool_inclusion_status import MempoolInclusionStatus
-from chia.util.condition_tools import conditions_dict_for_solution, pkm_pairs_for_conditions_dict
 from chia.util.errors import Err
 from chia.util.streamable import VersionedBlob
 from chia.wallet.puzzles.clawback.drivers import (
